Added implementation for basic end-to-end search from REST API to Solr. Added logging libraries and config, various tidying up of APIs

+++ src/main/scala/com/blinkboxbooks/common/spray/BlinkboxHelpers.scala
@@ -50,7 +50,7 @@ trait BlinkboxHelpers {
   /**
    * Generate links for use in paged results.
    */
-  def links(numberOfResults: Int, offset: Int, count: Int, linkBaseUrl: String) = {
+  def links(numberOfResults: Long, offset: Long, count: Long, linkBaseUrl: String) = {
 
     val hasMore = numberOfResults > offset + count
